fix: Make --inspect-port=0 work
Selecting port zero makes the inspector bind to a random port.

The inspector prints the URL it listens on to stderr.  Parse the
output from the child process to find the port number.

This commit coincidentally also makes `node inspect --port=0 x.js`
work and a regression test to that effect has been added.

Note that connecting to an existing process that listens on a
non-standard port with `node inspect -p <pid>` does not work
because then we never see the output from the inspector.

Fixes: https://github.com/nodejs/node/issues/16469