fix: Fix `takeHeapSnapshot()` truncation bug
Fix a logic bug in `takeHeapSnapshot()` where it interpreted the item
count as the byte count, resulting in truncated snapshots.

This change hinges on the observation that the completion callback
always comes after any and all `'addHeapSnapshotChunk'` events.

I'm 95% sure after digging into the V8 inspector and its integration
with Node.js that the assumption above is correct.

If it turns out I'm mistaken, then we will most likely treat that as
a bug in Node.js, not node-inspect.

Fixes: https://github.com/nodejs/node-inspect/issues/56