refactor: remove unused timeout parameter in portIsFree
`timeout` is specified as an optional parameter with a default value of
`9999`. `portIsFree` doesn't have any callsites that use this optional
parameter, so this removes the parameter and leaves the timeout value as
a local variable in the function. `portIsFree` isn't exported outside of
`_inspect.js`, so this doesn't change any public function interfaces.